ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pillow
python-multipart
rembg
opencv-python
onnxruntime